packages/gaic-widget/pyproject.toml
[project]
name = "gaic-widget"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anywidget[dev]>=0.9.13",
    "databricks-sdk>=0.36.0",
    "pydantic-settings>=2.6.0",
    "pydantic>=2.9.2",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5"
]

[tool.hatch.build.targets.wheel]
packages = ["src/gaic_widget"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"